크롤러 수정

+++ src/main/java/com/java/JustThree/repository/WebtoonRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WebtoonRepository extends JpaRepository<Webtoon,Integer> {
+public interface WebtoonRepository extends JpaRepository<Webtoon,Long> {
 
 }
